rename sparse_matrix2d
