+++ source/consumers/dataspaces/SparseMatrix2D.cpp
@@ -1,4 +1,4 @@
-#include <consumers/dataspaces/sparse_matrix2d.h>
+#include <consumers/dataspaces/SparseMatrix2D.h>
 #include <core/util/ascii_tree.h>
 #include <core/util/h5json.h>
