+++ tests/consumers/dataspaces/CMakeLists.txt
@@ -5,7 +5,7 @@ set(SOURCES
   ${dir}/scalar.cpp
   ${dir}/sparse_map2d.cpp
   ${dir}/sparse_map3d.cpp
-  ${dir}/sparse_matrix2d.cpp
+  ${dir}/SparseMatrix2DTest.cpp
   )
 
 set(${this_target}_sources ${${this_target}_sources} ${SOURCES} PARENT_SCOPE)
